flink/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
apache-beam==2.48.0
apache-flink==1.20.0
apache-flink-libraries==1.20.0
avro-python3==1.10.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
confluent-kafka==2.7.0
crcmod==1.7
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
exceptiongroup==1.2.2
fastapi==0.115.6
fastavro==1.10.0
fasteners==0.19
find_libpython==0.4.0
grpcio==1.68.1
h11==0.14.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
kafka-python==2.0.2
numpy==1.24.4
objsize==0.6.1
orjson==3.10.13
pandas==2.2.3
pemja==0.4.1
proto-plus==1.25.0
protobuf==4.23.4
py4j==0.10.9.7
pyarrow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.8
ruamel.yaml.clib==0.2.12
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0